Move Protocol3::Eeprom classes to base TimexDatalinkClient class.
